# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "2.0.0",
    "title": "Movie information API",
    "description": "This API provides access to a catalogue of movies, the people who made them, and ratings from several sources, together with JWT-based user accounts.\n\n**Note:** the original IMDB-derived dataset was hosted on PlanetScale, whose free tier shut down in April 2024. The API now ships with a small curated sample catalogue in SQLite so it runs with no external services. Titles, years and credits are real; ratings and box-office figures are approximate."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    },
    "schemas": {
...
      "TokenExpired": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "JWT token has expired"
          }
        }
      },
...
      "Forbidden": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Forbidden"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/user/{email}/profile": {
      "get": {
        "description": "Provide profile information",
        "tags": [
          "Authentication"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The profile's email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An object containing the profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "cyrus@gmail.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Cyrus"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Nguyen"
                    },
                    "dob": {
                      "type": "string",
                      "example": "2000-01-01"
                    },
                    "address": {
                      "type": "string",
                      "example": "Brisbane, Australia"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid refresh request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete, refresh token required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExpired"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "200 Not Owner": {
            "description": "An object containing the profile without dob and address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "cyrus@gmail.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Cyrus"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Nguyen"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Update profile information",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "description": "An object containing the first name (firstName), last name (lastName), date of birth (dob) and address (address).",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "example": "Cyrus"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Nguyen"
                  },
                  "dob": {
                    "type": "string",
                    "example": "2000-01-01"
                  },
                  "address": {
                    "type": "string",
                    "example": "Brisbane, Australia"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The profile's email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "An object containing the updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "cyrus@gmail.com"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Cyrus"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Nguyen"
                    },
                    "dob": {
                      "type": "string",
                      "example": "2000-01-01"
                    },
                    "address": {
                      "type": "string",
                      "example": "Brisbane, Australia"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Object does not contain all of the fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete: firstName, lastName, dob and address are required."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExpired"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated as a different user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "Profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "400 Not String": {
            "description": "Any of the fields are not strings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body invalid: firstName, lastName and address must be strings only."
                    }
                  }
                }
              }
            }
          },
          "400 Wrong Date Format": {
            "description": "If the date of birth is not a valid YYYY-MM-DD date (e.g. no April 31 or February 30, or February 29 on a non-leap year) or the date is not in the past",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid input: dob must be a real date in format YYYY-MM-DD."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "/",
      "description": "Same origin as these docs"
    }
  ]
}